data validation

+++ Requiem.PetProject/Controllers/CommentController.cs
@@ -21,6 +21,9 @@ public class CommentController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var comments = await _commentRepository.GetAllAsync();
 
         var commentDto = comments.Select(c => c.ToCommentDto());
@@ -28,59 +31,64 @@ public class CommentController : ControllerBase
         return Ok(commentDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:guid}")]
     public async Task<IActionResult> GetById([FromRoute] Guid id)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var comment = await _commentRepository.GetByIdAsync(id);
 
-        if (comment == null)
-        {
+        if (comment == null) 
             return NotFound();
-        }
 
         return Ok(comment.ToCommentDto());
     }
 
-    [HttpPost("{stockId}")]
+    [HttpPost]
+    [Route(("{stockId:guid}"))]
     public async Task<IActionResult> Create([FromRoute] Guid stockId, CreateCommentDto commentDto)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         if (!await _stockRepository.StockExists(stockId))
-        {
             return BadRequest("Stock does not exist");
-        }
 
         var commentModel = commentDto.ToCommentFromCreateDto(stockId);
 
         await _commentRepository.CreateAsync(commentModel);
 
-        return CreatedAtAction(nameof(GetById), new {id = commentModel}, commentModel.ToCommentDto());
+        return CreatedAtAction(nameof(GetById), new {id = commentModel.Id}, commentModel.ToCommentDto());
     }
 
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:guid}")]
     public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] UpdateCommentRequestDto updateDto)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var comment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdateDto(id));
 
-        if (comment == null)
-        {
+        if (comment == null) 
             return NotFound("Comment not found");
-        }
 
         return Ok(comment.ToCommentDto());
     }
     
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:guid}")]
     public async Task<IActionResult> Delete([FromRoute] Guid id)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var commentModel = await _commentRepository.DeleteAsync(id);
 
-        if (commentModel == null)
-        {
+        if (commentModel == null) 
             return NotFound("Comment does not exist");
-        }
-
+        
         return Ok(commentModel);
     } 
 }
