+++ Requiem.PetProject/Controllers/StockController.cs
@@ -1,6 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using Requiem.PetProject.Data;
 using Requiem.PetProject.DTOs.Stock;
 using Requiem.PetProject.Interfaces;
 using Requiem.PetProject.Mappers;
@@ -13,7 +11,7 @@ public class StockController : ControllerBase
 {
     private readonly IStockRepository _stockRepository;
     
-    public StockController(ApplicationDbContext context, IStockRepository stockRepository)
+    public StockController(IStockRepository stockRepository)
     {
         _stockRepository = stockRepository;
     }
@@ -21,21 +19,25 @@ public class StockController : ControllerBase
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var stocks = await _stockRepository.GetAllAsync();
         var stockDto = stocks.Select(s => s.ToStockDto());
 
         return Ok(stockDto);
     }
 
-    [HttpGet("{id}")]
+    [HttpGet("{id:guid}")]
     public async Task<IActionResult> GetById([FromRoute] Guid id)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var stock = await _stockRepository.GetByIdAsync(id);
 
         if (stock == null)
-        {
             return NotFound();
-        }
         
         return Ok(stock.ToStockDto());
     }
@@ -43,6 +45,9 @@ public class StockController : ControllerBase
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateStockRequestDto createDto)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var stockModel = createDto.ToStockFromCreateDto();
         await _stockRepository.CreateAsync(stockModel);
         
@@ -50,29 +55,31 @@ public class StockController : ControllerBase
     }
 
     [HttpPut]
-    [Route("{id}")]
+    [Route("{id:guid}")]
     public async Task<IActionResult> Update([FromRoute] Guid id, [FromBody] UpdateStockRequestDto updateDto)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var stockModel = await _stockRepository.UpdateAsync(id, updateDto);
 
         if (stockModel == null)
-        {
             return NotFound();
-        }
-
+        
         return Ok(stockModel.ToStockDto());
     }
 
     [HttpDelete]
-    [Route("{id}")]
+    [Route("{id:guid}")]
     public async Task<IActionResult> Delete([FromRoute] Guid id)
     {
+        if (!ModelState.IsValid) 
+            return BadRequest(ModelState);
+        
         var stockModel = await _stockRepository.DeleteAsync(id);
 
         if (stockModel == null)
-        {
             return NotFound();
-        }
 
         return NoContent();
     }
